+++ src/adjusters/ipv4.mjs
@@ -7,12 +7,7 @@ import Type from "../decorators/string/type";
 import Trim from "../decorators/string/trim";
 import Pattern from "../decorators/string/pattern";
 
-const PATTERN_COMPONENT = `(25[0-5]|2[0-4][0-9]|[0-1]?[0-9]{1,2})`;
-const PATTERN = `${PATTERN_COMPONENT}(\\.${PATTERN_COMPONENT}){3}`;
-
-const REGEXP = new RegExp(`^${PATTERN}$`);
-
-export {PATTERN};
+import {REGEXP} from "../libs/regexp/ipv4";
 
 /**
  * factory